tests: merge privateRepository.test into getPackageJson.test
